Provisioning: Run msvc 2019 update at the beginning of provisioning
When script was run with prefix 99 the installation process wasn't
finished before machine got rebooted. Moving the script at
beginning of provisioning seems to help with the situation. This is
not a proper fix but the best we have right now.
We can change this later if/when we find the root cause for this.

Task-number: QTQAINFRA-4813
Change-Id: Icbc2b0b33905990df611bdb494ed9b22801633fd
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>
(cherry picked from commit 1597f435123dfffbc2d7c14cc0a730d65c60387a)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>