Separate qtsaferenderer platforms from default platforms
Qtsaferenderer has mostly unique platforms and sharing them with
the default platforms makes the module_config.yaml condition for
the platforms impossible to maintain.

By separating the configs completely the conditions related to
the platforms can be deleted in the module_config.yaml. This
allows much simpler maintaining in single repo instead of
multiple repos.

Change-Id: Ia2d6d04536a21982b3da9c5e5ae77b783026a898
Reviewed-by: Antti Kokko <antti.kokko@qt.io>